requirements.txt
flask>=3.0.0
playwright>=1.40.0
schedule>=1.2.0
python-dotenv>=1.0.1
requests>=2.31.0
firebase-admin>=6.5.0
flask-limiter>=3.5.0
